+++ src/raport_slotow/models/uczelnia.py
@@ -20,9 +20,10 @@ from django.urls import reverse
 
 from bpp.core import zbieraj_sloty
 from bpp.fields import YearField
-from bpp.models import Cache_Punktacja_Autora_Query, Uczelnia
+from bpp.models import Autor, Cache_Punktacja_Autora_Query, Uczelnia
 from bpp.util import year_last_month
 from notifications.core import send_redirect
+from raport_slotow.core import autorzy_zerowi
 from raport_slotow.models.asgi_notification_mixin import ASGINotificationMixin
 from raport_slotow.models.mixins import Report
 
@@ -38,6 +39,13 @@ class RaportSlotowUczelnia(ASGINotificationMixin, Report):
     dziel_na_jednostki_i_wydzialy = models.BooleanField(
         verbose_name="Dziel na jednostki i wydziały",
         default=True,
+    )
+
+    pokazuj_zerowych = models.BooleanField(
+        verbose_name="Dołączaj autorów z zerowymi slotami",
+        default=False,
+        help_text="""Autor z zerowym slotem to autor, który ma na zadany okres czasu zadeklarowane dyscypliny,
+        ale nie posiada żadnych punktowanych prac. """,
     )
 
     def get_absolute_url(self):
@@ -103,6 +111,49 @@ class RaportSlotowUczelnia(ASGINotificationMixin, Report):
 
             if not n % 10:
                 self.send_progress(n * 100.0 / total)
+
+        if self.pokazuj_zerowych:
+            zerowi = autorzy_zerowi(
+                od_roku=self.od_roku, do_roku=self.do_roku, min_pk=self.minimalny_pk
+            )
+
+            # --- Początek --- komentarz dot. buga w Django
+            # Z uwagi na bug w django (obecnie 3.0.11) nie mozemy zrobic tutaj
+            # zerowi = zerowi.values_list("autor_id", "dyscyplina_naukowa_id").distinct()
+            # a ja na ten (29. sty 2021) moment nie mam czasu poprawiać Django, zatem uzyjemy
+            # set() aby lokalizować autorów już widzianych
+            # --- Koniec --- komentarza dot. buga w Django
+
+            # Jeżeli w raporcie jest włączony podział na jednostki i wydziały, to dla każdego autora zerowego
+            # dorzuć rekordy jego jednostek; jeżeli nie - to wrzuć po prostu autorów:
+
+            seen = set()
+
+            for autor_id, rok, dyscyplina_id in zerowi.values_list():
+                if (autor_id, dyscyplina_id) in seen:
+                    continue
+
+                seen.add((autor_id, dyscyplina_id))
+
+                kw = dict(
+                    autor_id=autor_id,
+                    jednostka_id=None,
+                    dyscyplina_id=dyscyplina_id,
+                    slot=0,
+                    pkd_aut_sum=0,
+                    avg=None,
+                )
+
+                if self.dziel_na_jednostki_i_wydzialy:
+                    for (jednostka_id,) in (
+                        Autor.objects.get(pk=autor_id)
+                        .autor_jednostka_set.values_list("jednostka_id")
+                        .distinct()
+                    ):
+                        kw["jednostka_id"] = jednostka_id
+                        self.raportslotowuczelniawiersz_set.create(**kw)
+                else:
+                    self.raportslotowuczelniawiersz_set.create(**kw)
 
     def on_finished_successfully(self):
         send_redirect(str(self.pk), "./details")