+++ src/raport_slotow/views/zerowy.py
@@ -1,20 +1,18 @@
 import urllib
 
+from django.contrib.postgres.aggregates.general import StringAgg
 from django.db import connection
 from django.db.models.fields import TextField
 from django.db.models.functions import Cast
 from django_filters.views import FilterView
 from django_tables2 import SingleTableMixin
+
+from bpp.views.mixins import UczelniaSettingRequiredMixin
+from raport_slotow.core import autorzy_zerowi
 from raport_slotow.filters import RaportZerowyFilter
 from raport_slotow.models import RaportZerowyEntry
 from raport_slotow.tables import RaportSlotowZerowyTable
 from raport_slotow.util import MyExportMixin, create_temporary_table_as
-
-from django.contrib.postgres.aggregates.general import StringAgg
-
-from bpp.models import Cache_Punktacja_Autora_Query_View
-from bpp.models.dyscyplina_naukowa import Autor_Dyscyplina
-from bpp.views.mixins import UczelniaSettingRequiredMixin
 
 
 class RaportSlotowZerowy(
@@ -40,26 +38,8 @@ class RaportSlotowZerowy(
         return context
 
     def get_queryset(self):
-        # wartośći zadeklarowane w bazie danych
-        defined = (
-            Autor_Dyscyplina.objects.values("autor_id", "rok", "dyscyplina_naukowa_id")
-            .exclude(dyscyplina_naukowa_id=None)
-            .union(
-                Autor_Dyscyplina.objects.values(
-                    "autor_id", "rok", "subdyscyplina_naukowa_id"
-                ).exclude(subdyscyplina_naukowa_id=None)
-            )
-        )
+        res = autorzy_zerowi(min_pk=self.min_pk)
 
-        # zestawy autor/rok/dyscyplina z całej bazy danych
-        existent = Cache_Punktacja_Autora_Query_View.objects.all()
-
-        if self.min_pk:
-            existent = existent.exclude(rekord__punkty_kbn__lte=self.min_pk)
-
-        existent = existent.values("autor_id", "rekord__rok", "dyscyplina_id")
-
-        res = defined.difference(existent)
         create_temporary_table_as("raport_slotow_raportzerowyentry", res)
         with connection.cursor() as cursor:
             cursor.execute(